06-lab-complete/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.0.1
langchain-openai==0.2.14
langchain-core==0.3.30
langgraph==0.2.62
typing-extensions==4.12.2
pydantic==2.10.6
